pyproject.toml
[project]
name = "cucbot-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "motor>=3.7.1",
    "python-dotenv>=1.2.1",
]
